+++ lorien/InfiniteCanvas/PanZoomCamera.gd
@@ -3,9 +3,11 @@ extends Camera2D
 signal zoom_changed(value)
 signal position_changed(value)
 
-const MAX_MOUSE_WHEEL_LEVEL := 1000
-const MIN_MOUSE_WHEEL_LEVEL := -9
+const MAX_MOUSE_WHEEL_LEVEL := 99
+const MIN_MOUSE_WHEEL_LEVEL := -8
 const ZOOM_INCREMENT = 0.1
+
+export var zoom_curve: Curve
 
 var _current_zoom_level := 1.0
 var _pan_active := false
@@ -15,7 +17,22 @@ var _mouse_wheel_counter := 0
 
 # -------------------------------------------------------------------------------------------------
 func set_zoom_level(zoom_level: float) -> void:
-	_current_zoom_level = zoom_level
+	_current_zoom_level = 1.0
+	_mouse_wheel_counter = 0
+	
+	if !is_equal_approx(zoom_level, 1.0):
+		var wheel_direction := -1 if (zoom_level < 1.0) else 1
+		
+		while true:
+			_mouse_wheel_counter += wheel_direction
+			_current_zoom_level = 1.0 + _mouse_wheel_counter * _calc_zoom_increment()
+			
+			if wheel_direction < 0:
+				if _current_zoom_level <= zoom_level:
+					break
+			else:
+				if _current_zoom_level >= zoom_level:
+					break
 	zoom = Vector2(_current_zoom_level, _current_zoom_level)
 
 # -------------------------------------------------------------------------------------------------
@@ -63,19 +80,10 @@ func _do_zoom() -> void:
 
 # -------------------------------------------------------------------------------------------------
 func _calc_zoom_increment() -> float:
-	var incr := ZOOM_INCREMENT
-	if _mouse_wheel_counter >= 100:
-		incr *= 5.0
-	elif _mouse_wheel_counter >= 75:
-		incr *= 4.0
-	elif _mouse_wheel_counter >= 50:
-		incr *= 3.0
-	elif _mouse_wheel_counter >= 25:
-		incr *= 2.0
-	elif _mouse_wheel_counter >= 10:
-		incr *= 1.5
-	return incr
-	
+	var progress: float = _mouse_wheel_counter + abs(MIN_MOUSE_WHEEL_LEVEL)
+	progress /= (abs(MIN_MOUSE_WHEEL_LEVEL) + MAX_MOUSE_WHEEL_LEVEL)
+	return zoom_curve.interpolate(progress)
+
 # -------------------------------------------------------------------------------------------------
 func enable_intput() -> void:
 	_is_input_enabled = true